refactor: refactored conference management to use new file and folder structure
