+++ src/app/api/transform/conference.0/route.ts
@@ -1,57 +1,13 @@
-import { eq } from "drizzle-orm"
-
-import {
-  conference,
-  ConferenceEventArchivedPayload,
-  ConferenceEventCreatedPayload,
-  ConferenceEventUpdatedPayload,
-} from "@/contracts/events/conference"
-import { db } from "@/database"
-import { conferences } from "@/database/schemas"
+import conferenceArchived from "@/app/api/transform/conference.0/route-conference-archived"
+import conferenceCreated from "@/app/api/transform/conference.0/route-conference-created"
+import conferenceUpdated from "@/app/api/transform/conference.0/route-conference-updated"
+import { conference } from "@/contracts/events/conference"
 import EventTransformer from "@/lib/event-transformer"
 
 const eventTransformer = new EventTransformer(conference, {
-  created: async (payload: unknown) => {
-    console.log("Got created event", payload)
-    const parsedPayload = ConferenceEventCreatedPayload.parse(payload)
-    const exists = await db.query.conferences.findFirst({
-      where: eq(conferences.id, parsedPayload.id),
-    })
-    if (exists) {
-      return
-    }
-    await db.insert(conferences).values(parsedPayload)
-  },
-  updated: async (payload: unknown) => {
-    console.log("Got updated event", payload)
-    const parsedPayload = ConferenceEventUpdatedPayload.parse(payload)
-    const exists = await db.query.conferences.findFirst({
-      where: eq(conferences.id, parsedPayload.id),
-    })
-    if (!exists) {
-      return
-    }
-    await db
-      .update(conferences)
-      .set(parsedPayload)
-      .where(eq(conferences.id, parsedPayload.id))
-  },
-  archived: async (payload: unknown) => {
-    console.log("Got archived event", payload)
-    const parsedPayload = ConferenceEventArchivedPayload.parse(payload)
-    const exists = await db.query.conferences.findFirst({
-      where: eq(conferences.id, parsedPayload.id),
-    })
-    if (!exists) {
-      return
-    }
-    await db
-      .update(conferences)
-      .set({
-        archived: true,
-      })
-      .where(eq(conferences.id, parsedPayload.id))
-  },
+  created: conferenceCreated,
+  updated: conferenceUpdated,
+  archived: conferenceArchived,
 })
 
 export const POST = eventTransformer.post.bind(eventTransformer)
