+++ src/server/api/root.ts
@@ -1,4 +1,4 @@
-import { conferenceRouter } from "@/server/api/routers/conferences"
+import { conferenceRouter } from "@/server/api/routers/conference/conference.router"
 import { profileRouter } from "@/server/api/routers/profile"
 import ticketRouter from "@/server/api/routers/ticket/ticket.router"
 import { userRouter } from "@/server/api/routers/user"
